backend/requirements.txt
# SnackScan Backend Requirements
# Core Flask application
Flask==3.0.0
Flask-CORS==4.0.0

# Image processing and computer vision
Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.3

# Google Vision API (optional - for advanced image recognition)
google-cloud-vision==3.4.5

# HTTP requests for nutrition APIs
requests==2.31.0

# Environment variable management
python-dotenv==1.0.0

# JSON handling and data processing
pydantic==2.5.0

# Logging and utilities
colorlog==6.8.0

# Development and testing (optional)
pytest==7.4.3
pytest-flask==1.3.0
black==23.11.0
flake8==6.1.0

# Production WSGI server
gunicorn==21.2.0